requirements.txt
Django==5.1.4
djangorestframework==3.15.2
psycopg2-binary==2.9.10
requests==2.32.3
python-dotenv==1.0.1
drf-spectacular==0.27.2
# Testing & Linting
pytest==8.3.4
pytest-django==4.9.0
ruff==0.8.4
factory-boy==3.3.1